Setup Redux and Asynchronous API interactions:
Added Redux to manage state. I've also added a number of dependencies
to provide the ability to call remote APIs, as well as some debugging
tools to make Redux development easier.

Added code to index.js to setup the middleware (thunk) to be used to
create async action creators used to make requests of remote APIs. It
also adds the previously mentioned development tools (in `development`
mode only), these are all used to create the Redux store used in the
app.