Provide the ability to combing multiple reducers:
I anticipate the need for more reducers in the future, so got the
problem out of the way now (just as I'm adding my second reducer).
I have tested this with another reducer but am not ready to commit
it yet, so I'm just committing the ability to combine reducers, I'll
add the next reducer when I commit the next feature.