Add Generic Form validation:
Created a Generic Form validation system that can be used against any
form. It uses a Higher Order Component (HOC) function to map state and
validation rules to a Stateless Functional Component that in this case
is the LoginForm.

This solution was found at:
https://medium.com/javascript-inside/form-validation-as-a-higher-order-component-pt-2-1edb7881870d

I've tweaked it a bit, mainly splitting up concerns a bit more and moving
from data.either to folktale.

I've removed the connecting of the form with Redux, so submitting no
longer works (in fact there is a bug at the moment with the onSubmit
function I've implememented, I'll be fixing that next, then re-adding
the connection to Redux.