Re-introduce form submission:
When adding the form validation, I temporarily disabled this aspect,
I have now re-introduced it. I'm not yet sure this is the best solution
but it is a, solution.

I've noticed that there is no user in the login state, after having
logged in, need to check if that was the case in the master branch
or have I broken the reducer.