Make Map component grow in size:
Rather than had-coding the map div element's size in pixels, I set
the width/height of the page body and set the map to grow in size.

I replaced the bootstrap column style for one of my own that uses flex
to grow the element to take up the available space i.e. `flex: 1;`.

I also removed the unnecessary div element from around the `<Map />`
element.

NOTE: I also changed the CSS for the map to use viewport unit `vh`
to set the map height. I tried `100%`, but that placed the map over
the `<Footer />`.